refactor(svelte): extract product chrome to private @epicenter/app-shell
AccountPopover and WorkspaceGate are branded, full-stack-aware app chrome
(they pull @epicenter/ui, @lucide/svelte, and @tanstack/svelte-query and
hardcode Epicenter copy). They do not belong in @epicenter/svelte, which is
slated to be published as the framework reactive-bindings layer.

Move both components to a new private @epicenter/app-shell package, drop the
two chrome subpath exports plus the now-chrome-only deps from @epicenter/svelte,
and re-point the five consuming apps. @epicenter/svelte is now pure .svelte.ts
reactive modules over auth + workspace.
