+++ apps/honeycrisp/package.json
@@ -35,6 +35,7 @@
 		"vite": "catalog:"
 	},
 	"dependencies": {
+		"@epicenter/app-shell": "workspace:*",
 		"@epicenter/auth": "workspace:*",
 		"@epicenter/field": "workspace:*",
 		"@epicenter/svelte": "workspace:*",
