+++ apps/opensidian/package.json
@@ -22,6 +22,7 @@
 	},
 	"dependencies": {
 		"@codemirror/autocomplete": "^6.20.1",
+		"@epicenter/app-shell": "workspace:*",
 		"@epicenter/auth": "workspace:*",
 		"@epicenter/constants": "workspace:*",
 		"@epicenter/field": "workspace:*",
