+++ apps/tab-manager/package.json
@@ -20,6 +20,7 @@
 		"postinstall": "wxt prepare"
 	},
 	"dependencies": {
+		"@epicenter/app-shell": "workspace:*",
 		"@epicenter/auth": "workspace:*",
 		"@epicenter/constants": "workspace:*",
 		"@epicenter/field": "workspace:*",
