+++ apps/zhongwen/package.json
@@ -20,6 +20,7 @@
 		"deploy": "wrangler deploy"
 	},
 	"dependencies": {
+		"@epicenter/app-shell": "workspace:*",
 		"@epicenter/auth": "workspace:*",
 		"@epicenter/constants": "workspace:*",
 		"@epicenter/field": "workspace:*",
