+++ packages/app-shell/src/account-popover/account-popover.svelte
