+++ packages/app-shell/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "../../tsconfig.dom.json",
+	"compilerOptions": {
+		"types": ["bun"],
+		"noUnusedLocals": true,
+		"noUnusedParameters": true
+	}
+}
